refactor: Move dbTypeIndexRefersTo ImportIndexDto to  #10077
